+++ CHANGELOG.md
@@ -60,3 +60,7 @@
 * Define `Or` data type
 * Define `U_I_UU_I_II` data type
 * Define `Constant` data type
+* Define `Void` data type
+* Define `Unit` data type
+* Define `Representation` type family
+* Define `Representable` type alias